feat: 更加合理的数据库初始化流程
